There are three categories of tests:
1. Conformance tests: inside ../should_pass/
These tests verify the support and the correctness of the
implementation of features that DONOT require multiple runs.
examples include support of coarrays, cosubscripts, CODIMENSION
attribute, etc.

2. Confidence tests: inside ../confidence_tests
These tests use cross-testing to verify correctness. The tests
verify the correctness of constructus that are used to handle
the non-determinism that arises die to one-sided SPMD style
model. Every test for a construct is tested against a version
without the construct. The level of confidence in the tests is
displayed for every test. Note that to gain high confidence, it
is essential that the tests be run for large number of images and
iterations.

3. Faulty tests: inside ../faulty_tests
These tests check for the correctness of support in the standard
to handle special/ faulty conditions (e.g. images waiting at a
barrier for others which have already terminated). (e.g. use of
STAT specifiers).

4. Specific tests: inside ../few_tests
These tests only those tests which have been manually added in
the user-file called 'test_file'. This is helpful when users
want to test very specific tests which fall under one of the
aboe category. It must be noted, that the method of testing of
eack of the able categories and this testing infrasructure
associates every listed test with one of the above categories.

Other features added:
1. Augmentation of existing tests to verify support of both
ALLOCATABLES and SAVE coarrays for certains
constructs/intrinsics.
2. Tests with  REAL and INTEGER data types.
3. Tests with  strided cosubscripts.
4. Made the tests independent of the location of installation.

TESTING:
Tested with OpenUH & Intel compiler.

TODO:
1. for certain cases when the relationship between the cores and
the processes is one-to-many, processes stall and then timesout.
During these instances, the perl script timedexec.pl doesnot
terminate the processes. This needs to be looked into.
2. Some tests could also benefit from testing with different
sized INTEGER and REAL coarrays.
3. Test with other compilers e.g. GNU, Cray

Contributor: Siddhartha